feat: invitation & content base setup

+++ db/migrate/20250710103230_add_dob_and_parental_consent_to_users.rb
@@ -2,24 +2,17 @@ class AddDobAndParentalConsentToUsers < ActiveRecord::Migration[8.0]
   def up
     return unless table_exists?(:users)
 
-    unless column_exists?(:users, :date_of_birth)
-      add_column :users, :date_of_birth, :date
-    end
-
-    unless column_exists?(:users, :parental_consent)
-      add_column :users, :parental_consent, :boolean, default: false, null: false
-    end
+    add_column :users, :date_of_birth, :date unless column_exists?(:users, :date_of_birth)
+    add_column :users, :first_name, :string  unless column_exists?(:users, :first_name)
+    add_column :users, :parental_consent, :boolean, default: false, null: false unless column_exists?(:users, :parental_consent)
   end
 
   def down
     return unless table_exists?(:users)
 
-    if column_exists?(:users, :date_of_birth)
-      remove_column :users, :date_of_birth
-    end
-
-    if column_exists?(:users, :parental_consent)
-      remove_column :users, :parental_consent
-    end
+    remove_column :users, :date_of_birth if column_exists?(:users, :date_of_birth)
+    remove_column :users, :first_name if column_exists?(:users, :first_name)
+    remove_column :users, :last_name if column_exists?(:users, :last_name)
+    remove_column :users, :parental_consent if column_exists?(:users, :parental_consent)
   end
 end
